DELIA-51403: [AAMP][DASH][PEACOCK]Content Audio Played for 0-1 seconds After AdUI started, Fix for negative buffer
Reason for change: Incrementing priority of period duration calculation
based on period start times. This change allows fragment collector to skip fragments
beyond next period start time. Also added Offset delta to start time and fixed negative buffer
valuses
Risks: Low
Test Procedure: Test with STBs.

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_21_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_21_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_21_1

+++ fragmentcollector_mpd.h
@@ -48,7 +48,7 @@ using namespace dash::helpers;
 /*Common MPD util functions (admanager_mpd.cpp and fragmentcollector_mpd.cpp */
 uint64_t aamp_GetPeriodNewContentDuration(IPeriod * period, uint64_t &curEndNumber);
 double aamp_GetPeriodStartTimeDeltaRelativeToPTSOffset(IPeriod * period);
-uint64_t aamp_GetPeriodDuration(dash::mpd::IMPD *mpd, int periodIndex, uint64_t mpdDownloadTime = 0);
+double aamp_GetPeriodDuration(dash::mpd::IMPD *mpd, int periodIndex, uint64_t mpdDownloadTime = 0);
 Node* aamp_ProcessNode(xmlTextReaderPtr *reader, std::string url, bool isAd = false);
 uint64_t aamp_GetDurationFromRepresentation(dash::mpd::IMPD *mpd);
